# apifox.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "实时协作任务看板 API",
    "description": "Next.js App Router + Prisma 7 + NextAuth v5 实现的项目看板后端接口。所有接口基于 Cookie 会话（NextAuth）鉴权，统一响应格式为 { success, data?, error? }。看板列为固定的四种任务状态（TaskState）：TODO=未开始、UP_NEXT=接下来、IN_PROGRESS=进行中、DONE=已完成。",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "本地开发环境"
    }
  ],
...
  "security": [
    { "sessionCookie": [] }
  ],
  "paths": {
...
    "/api/projects/{id}": {
      "parameters": [
        { "$ref": "#/components/parameters/ProjectId" }
      ],
      "get": {
        "tags": ["Projects"],
        "summary": "获取项目详情",
        "description": "返回项目基本信息、所有者、成员列表，以及当前用户在项目中的角色（role）、活动数量（activityCount）、项目级讨论的已读水位线（lastReadAt）与未读起点（firstUnreadCommentId）——后两者都只属于请求者本人，lastReadAt 为 null 表示从未读过。",
        "responses": {
          "200": {
            "description": "项目详情",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseProjectDetail" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "patch": {
        "tags": ["Projects"],
        "summary": "更新项目（仅 ADMIN）",
        "description": "更新项目名称或描述，至少提供一个字段。name 不能为空字符串；description 传 null 或空清除。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateProjectRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseProject" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "delete": {
        "tags": ["Projects"],
        "summary": "删除项目（仅 ADMIN）",
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseDeleted" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
    "/api/projects/{id}/members": {
      "parameters": [
        { "$ref": "#/components/parameters/ProjectId" }
      ],
      "get": {
        "tags": ["Members"],
        "summary": "获取项目成员列表",
        "description": "返回项目成员（含用户信息），按加入时间升序。",
        "responses": {
          "200": {
            "description": "成员列表",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseMemberList" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "post": {
        "tags": ["Members"],
        "summary": "添加项目成员（仅 ADMIN）",
        "description": "按邮箱邀请成员，role 非法或缺省时默认 MEMBER。用户不存在返回 404，已是成员返回 400。写入 ADD_MEMBER 活动并广播 member:added。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddMemberRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "添加成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseMember" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "delete": {
        "tags": ["Members"],
        "summary": "移除项目成员（仅 ADMIN）",
        "description": "通过查询参数 userId 指定要移除的成员，不能移除自己（400）。成员不存在返回 404。写入 REMOVE_MEMBER 活动并广播 member:removed。",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "description": "要移除的用户 ID，不能为当前用户自己",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "移除成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseDeleted" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
    "/api/projects/{id}/comments/read": {
      "parameters": [
        { "$ref": "#/components/parameters/ProjectId" }
      ],
      "post": {
        "tags": ["Comments"],
        "summary": "标记项目级讨论已读（推进未读水位线）",
        "description": "把当前成员在项目级讨论（cardId 为 null 的评论）里的已读水位线推进到指定时间：不传 lastReadAt 表示推进到最新一条评论的 createdAt（即全部已读）；水位线只允许前进，并且以最新一条评论的时间为上界（伪造未来时间无效）；项目内没有任何评论时水位线不动。幂等：值没有变化时不写库。未读判定在前端完成：未读 = createdAt 晚于水位线 且 未解决 且 不是自己发的。",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CommentReadRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "标记成功（返回推进后的水位线；无评论且从未读过时为 null）",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseCommentReadState" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "authjs.session-token",
        "description": "NextAuth v5 会话 Cookie。生产环境（HTTPS）下名称为 __Secure-authjs.session-token。请先通过 /api/auth 登录后由浏览器自动携带。"
      }
    },
    "parameters": {
      "ProjectId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "项目 ID（cuid）",
        "schema": { "type": "string" }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "请求参数错误",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiErrorResponse" },
            "example": { "success": false, "error": "缺少必要参数" }
          }
        }
      },
      "Unauthorized": {
        "description": "未登录",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiErrorResponse" },
            "example": { "success": false, "error": "未登录" }
          }
        }
      },
      "Forbidden": {
        "description": "无权限",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiErrorResponse" },
            "example": { "success": false, "error": "无权访问该项目" }
          }
        }
      },
      "NotFound": {
        "description": "资源不存在",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiErrorResponse" },
            "example": { "success": false, "error": "不存在" }
          }
        }
      },
      "ServerError": {
        "description": "服务器内部错误",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiErrorResponse" },
            "example": { "success": false, "error": "操作失败" }
          }
        }
      }
    },
    "schemas": {
...
      "MemberRole": {
        "type": "string",
        "enum": ["ADMIN", "MEMBER"],
        "description": "项目成员角色"
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string", "nullable": true },
          "email": { "type": "string", "nullable": true },
          "image": { "type": "string", "nullable": true }
        }
      },
      "UserBrief": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string", "nullable": true },
          "image": { "type": "string", "nullable": true }
        }
      },
...
      "Project": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "ownerId": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "ProjectDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/Project" },
          {
            "type": "object",
            "properties": {
              "role": { "$ref": "#/components/schemas/MemberRole" },
              "activityCount": { "type": "integer", "description": "项目活动记录总数" },
              "lastReadAt": {
                "type": "string",
                "format": "date-time",
                "nullable": true,
                "description": "当前请求者对项目级讨论的已读水位线（值为已读到的最后一条评论的 createdAt，null 表示从未读过）。只属于请求者本人，不随 members 下发"
              },
              "firstUnreadCommentId": {
                "type": "string",
                "nullable": true,
                "description": "打开项目时的未读起点：项目级讨论里第一条「晚于本人水位线、未解决、不是自己发的」评论 id（null 表示没有未读）。前端据此在第一条未读之前画「以下为未读消息」分割线，本次会话内固定不变"
              },
              "owner": { "$ref": "#/components/schemas/UserBrief" },
              "members": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/Member" }
              }
            }
          }
        ]
      },
      "Member": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "role": { "$ref": "#/components/schemas/MemberRole" },
          "joinedAt": { "type": "string", "format": "date-time" },
          "user": { "$ref": "#/components/schemas/User" }
        }
      },
...
      "CommentReadRequest": {
        "type": "object",
        "description": "标记项目级讨论已读；body 可以为空对象（等价于推进到最新一条评论）",
        "properties": {
          "lastReadAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "要推进到的水位线（ISO 8601，一般取已加载的最新一条评论的 createdAt）。缺省或 null 表示推进到最新一条评论；晚于最新评论时按最新评论截断；早于当前水位线时忽略（只前进不回退）",
            "example": "2026-03-01T08:30:00.000Z"
          }
        }
      },
...
      "UpdateProjectRequest": {
        "type": "object",
        "description": "至少提供一个字段",
        "properties": {
          "name": { "type": "string", "description": "项目名称，不能为空字符串" },
          "description": { "type": "string", "nullable": true, "description": "项目描述，传 null 或空清除" }
        }
      },
...
      "AddMemberRequest": {
        "type": "object",
        "required": ["email"],
        "properties": {
          "email": { "type": "string", "format": "email", "description": "要添加的成员邮箱，用户不存在返回 404" },
          "role": {
            "allOf": [{ "$ref": "#/components/schemas/MemberRole" }],
            "description": "成员角色，非法或缺省时为 MEMBER"
          }
        }
      },
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string" }
        }
      },
      "ApiResponseDeleted": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": {
            "type": "object",
            "properties": { "deleted": { "type": "boolean", "example": true } }
          }
        }
      },
...
      "ApiResponseProject": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": { "$ref": "#/components/schemas/Project" }
        }
      },
      "ApiResponseProjectDetail": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": { "$ref": "#/components/schemas/ProjectDetail" }
        }
      },
...
      "ApiResponseMember": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": { "$ref": "#/components/schemas/Member" }
        }
      },
      "ApiResponseMemberList": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Member" }
          }
        }
      },
...
      "ApiResponseCommentReadState": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": {
            "type": "object",
            "properties": {
              "lastReadAt": {
                "type": "string",
                "format": "date-time",
                "nullable": true,
                "description": "推进后的水位线；项目内没有评论时为 null"
              }
            }
          }
        }
      },
...
    }
  }
}